// src/app/core/interfaces/consulta.interface.ts
export interface Consulta {

  toca: string;

  recepcion: string;

  juzgadoProcedencia: string;

  expediente: string;

  dvd: string;

  conPreso: string;

  lugarCereso: string;

  resolucionApelada: string;

  fechaResolucion: string;

  imputado: string;

  delito: string;

  victima: string;

  apelante: string;

  fechaRadicacion: string;

  proyectista: string;

  sentidoResolucion: string;

  promueveAmparo: string;

  observaciones: string;

}